Accept a dict style of filter definition
Sliceable Switch REST API does not have a consistent API object model.
A network and a port are passed one by one. On the other hand a definition
of packet filter must be passed as a list.
It would be better that REST API accepts both a dict and a list.

This commit allows a user to post a dict style of packet filter definition
in addition to a list style.

A list support for network/port will be done in a later patch.