[sliceable_switch] fixed config.cgi to create wildcards properly
When the REST-API received a POST request to /filters without
wildcards, the wildcards are not created properly. The reason
is that unspecified parameters are defined as 0 at the oct()
iterpretation before set_default_filter_criteria() is called,
while wildcards are created by checking parameters are defined
or not. This patch fixed this by adding condition to the oct()
iterpretation so that unspecified parameters are kept undefined.

This patch also includes tests of REST-API to configure filters.