Use hex() instead of oct() in Sliceable Switch Filter.pm
oct() in Perl 5.18.2 on Ubuntu 14.04 does not accept a value
with non-octal integer for a case where the value is not intepreted
as hexdecimal. If a deciaml integer is passed to oct(), it is regarded
as a string and an expression like oct(4194303) results in NaN.
As a result, the default filter entry inserted by create_tables.sh
does not have a correct value. This completely Sliceable Switch
behavior completely because all traffic will be dropped.

This commit replaces oct() with hex().
hex() is more robust for input integers compared to oct().

Fixes #39